feat: root navigation tests + extract awala callings to a separate class to make AwalaManager testable + extract Logger to a separate class (#101)
I was trying to figure out what causes the issue from the task https://relaycorp.atlassian.net/browse/LTR-116. Have not found it in the code, but just to make sure that everything works as expected and we don't let a user to go to the Home screen before Awala is initialized, I wrote unit tests with different possible cases

So, in this PR:
1. Unit tests for root navigation screen (screen that is placed first in Letro's navigation stack)
2. Callings of Awala SDK's methods extracted to a wrapper class (`AwalaWrapper`) to make `AwalaManager` testable
3. Created `Logger` class to make non-Android classes testable (I refactored only classes that are used in the tests in this PR to make the PR simple, so I didn't change all `Log.method() `callings to the `Logger.method()` callings)

P.S. @gnarea I've checked that library for logging that you shared with me in some of your PRs. I tried to integrate it, but for some reasons, I did not see any logs when I used classes from this library. So I decided to not spend time on figure out the issue, and make a fast solution with my own logger wrapper